LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/origin-keyed-agent-clusters/1-iframe/parent-no-child-yes-same.sub.https-expected.txt


PASS setting document.domain must give sync access
FAIL parent: originAgentCluster must equal false assert_equals: expected (boolean) false but got (undefined) undefined
FAIL child: originAgentCluster must equal false assert_equals: expected (boolean) false but got (undefined) undefined
